requirements.txt
Cython==0.28.1
decorator==4.3.0
dload==0.6
funcsigs==1.0.2
gym==0.14.0
h5py==2.7.1
imageio==2.1.2
matplotlib==2.2.2
moviepy==0.2.3.2
numpy==1.18.4
opencv-python==3.4.2.17
Pillow==6.0.0
pyglet==1.2.4
scikit-image==0.15.0
scipy==1.3.0
tensorflow==1.14.0
tensorboardX==2.0
torch==1.1.0
torchvision==0.2.0
tqdm==4.11.2
urllib3==1.22
numpy-stl==2.11.2
pyquaternion==0.9.5
